requirements.txt
### bayts requirements
rioxarray>=.0.3.1
seaborn>=0.11.1
geopandas>=0.9.0
geocube>=-0.0.16
iteration_utilities>=-0.11.0
scipy>=1.6.1
sklearn
tqdm
fire
# rio-cogeo==2.0.1
# cogeo-mosaic==3.0.0rc1
# rio-tiler==2.0.3
# titiler==0.1.0
# attrs==20.3
